feat(client): delete field component
