+++ packages/amplication-client/src/Entity/EntityField.scss
@@ -1,0 +1,9 @@
+@import "../style/index.scss";
+
+.entity-field {
+  &__header {
+    @include flexFullRowWithSpacing;
+    justify-content: flex-end;
+    margin-bottom: var(--default-spacing);
+  }
+}
